fix "show crypto device mapping" cli command
If the optional parameter, "verbose", was not provided the display
function simply returned without any output to the terminal.

The unformat_input_t structure has no user command in the buffer when
optional parameters are not given.  As a result, unformat_user() returns
zero.  In this case, generate terse output.